change folder structure of login and sign up
