+++ src/App.js
@@ -1,13 +1,13 @@
 import "./App.css";
 import Home from "./Pages/Home";
-import Login from "./Pages/Login";
+import Form from "./Pages/UserForm/Form";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 const App = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Login />} />
+        <Route path="/" element={<Form />} />
         <Route path="/home" element={<Home />} />
       </Routes>
     </BrowserRouter>
