+++ src/Pages/UserForm/validate.js
@@ -3,18 +3,18 @@ export default function validate(values) {
 
   //   firstname
   if (!values.firstname) {
-    errors.firstname = "نام خانوادگی اجباری است !";
+    errors.firstname = "نام اجباری است !";
   }
 
   //   lastname
   if (!values.lastname) {
-    errors.lastname = "نام اجباری است !";
+    errors.lastname = "نام خانوادگی اجباری است !";
   }
 
   // username
-  //   if (!values.username) {
-  //     errors.username = "نام کاربری اجباری است !";
-  //   }
+  // if (!values.username) {
+  //   errors.username = "نام کاربری اجباری است !";
+  // }
 
   //   email
   if (!values.email) {
